상품 엔티티 및 주문 관련 낙관적 락 필드 추가, 재고 관리 로직 개선, 마이그레이션 스크립트 갱신, 테스트 수정

+++ boiler/src/main/kotlin/com/raynor/demo/boiler/controller/product/ProductController.kt
@@ -24,7 +24,7 @@ class ProductController(
     @GetMapping("")
     fun getProducts(
         @Min(1) @Max(100)
-        @RequestParam("size", required = false, defaultValue = "20") size: Long,
+        @RequestParam("size", required = false, defaultValue = "20") size: Int,
         @RequestParam("cursor", required = false) cursor: Int?,
     ): ResponseEntity<CursorPageResponseDto<Int, ProductResponseDto>> {
         return productService.getProducts(size, cursor).let { cursorSlice ->
